File storage max upload limit reform:
the max upload limit will now default to empty. Rationale: this value is
one that ages quickly with respect to new storage requirements, and is
currently set to something incompatible with the vanilla Naviserver configuration
(500MB fs vs 20MB ns).

The upgrade script will make so that instances setting this parameter
higher than the server upload limit configuration will be
set empty as well (this would never have an effect, as server
will prevent the upload altogheter anyway).

Max upload limit in bytes is now retrieved from a proc implementing
this fallback to server configuration logics. Other than centralizing,
its purpose is to be used in client-side code checks in the future
so we could prevent a doomed upload in advance.